sources: refactor deb source into module.
Extract the Deb source into `sources/_deb.py`.

LP: #1648906
