+++ snapcraft/internal/sources/__init__.py
@@ -81,7 +81,6 @@ import tempfile
 import tarfile
 import zipfile
 
-import apt_inst
 import libarchive
 
 from snapcraft.internal import common
@@ -89,6 +88,7 @@ from snapcraft import file_utils
 from . import errors
 from . import _base
 
+from ._deb import Deb  # noqa
 
 logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
@@ -399,39 +399,6 @@ class Zip(_base.FileBase):
 
         if not keep_zip:
             os.remove(zip)
-
-
-class Deb(_base.FileBase):
-
-    def __init__(self, source, source_dir, source_tag=None, source_commit=None,
-                 source_branch=None, source_depth=None):
-        super().__init__(source, source_dir, source_tag, source_commit,
-                         source_branch, source_depth)
-        if source_tag:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify a source-tag for a deb source')
-        elif source_commit:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify a source-commit for a deb source')
-        elif source_branch:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify a source-branch for a deb source')
-
-    def provision(self, dst, clean_target=True, keep_deb=False):
-        deb_file = os.path.join(self.source_dir, os.path.basename(self.source))
-
-        if clean_target:
-            tmp_deb = tempfile.NamedTemporaryFile().name
-            shutil.move(deb_file, tmp_deb)
-            shutil.rmtree(dst)
-            os.makedirs(dst)
-            shutil.move(tmp_deb, deb_file)
-
-        deb = apt_inst.DebFile(deb_file)
-        deb.data.extractall(dst)
-
-        if not keep_deb:
-            os.remove(deb_file)
 
 
 class Rpm(_base.FileBase):
